+++ data-transfer-object/src/main/java/org/grnet/status/dtos/tenant/node/WebApiNodeReportResponse.java
@@ -1,0 +1,13 @@
+package org.grnet.status.dtos.tenant.node;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.grnet.status.dtos.Status;
+
+@Setter
+@Getter
+public class WebApiNodeReportResponse {
+
+    private Status status;
+
+}