+++ data-transfer-object/src/main/java/org/grnet/status/dtos/tenant/node/WebApiNodeResponse.java
@@ -1,0 +1,16 @@
+package org.grnet.status.dtos.tenant.node;
+
+import org.grnet.status.dtos.Status;
+
+public class WebApiNodeResponse {
+
+    private Status status;
+
+    public Status getStatus() {
+        return status;
+    }
+
+    public void setStatus(Status status) {
+        this.status = status;
+    }
+}