+++ entity/src/main/resources/db/migration/tables/tenant/V1.14__update_tenant_v3.sql
@@ -1,0 +1,6 @@
+-- ------------------------------------------------
+-- Version: V1.14
+-- Description: Update t_tenant table to add status column
+-- ------------------------------------------------
+ALTER TABLE t_Tenant
+ADD COLUMN node BOOLEAN DEFAULT NULL;